Decompiled func_8000A184

+++ include/code0/24610.h
@@ -4,7 +4,7 @@
 #include "common.h"
 
 void func_80023A10(void);
-void func_80023C04(u16, u16, s32);
+void func_80023C04(s32, u16, s32);
 void func_80027C18(f32 x1, f32 y1, f32 x2, f32 y2, u16 tilenum, s32);
 void func_80028F04(u8 r1, u8 g1, u8 b1, u8 r2, u8 g2, u8 b2);
 void func_8002900C(u8 r1, u8 g1, u8 b1, u8 r2, u8 g2, u8 b2, u8 a);
